+++ lib/servy/pledge_server.ex
@@ -1,70 +1,50 @@
-defmodule Servy.GenericServer do
-
-  def start(callback_module, initial_state, name) do
-    pid = spawn(__MODULE__, :listen_loop, [initial_state, callback_module])
-    Process.register(pid, name)
-    pid
-  end
-
-  def call(pid, message) do
-    send(pid, {:call, self(), message})
-    receive do {:response, response} -> response end
-  end
-
-  def cast(pid, message) do
-    send(pid, {:cast, message})
-  end
-
-  def listen_loop(state, callback_module) do
-    receive do
-      {:call, sender, message} when is_pid(sender)->
-        {response, new_state} = callback_module.handle_call(message, state)
-        send(sender, {:response, response})
-        listen_loop(new_state, callback_module)
-
-      {:cast, message} ->
-        new_state = callback_module.handle_cast(message, state)
-        listen_loop(state, callback_module)
-
-      unexpected ->
-        IO.puts "Unexpected message #{inspect unexpected}"
-        listen_loop(state, callback_module)
-    end
-  end
-end
-
-
 defmodule Servy.PledgeServer do
-  alias Servy.GenericServer
 
   # this is used to register the process
   @name :pledge_server
 
-  # Client functions below ⬇️
+  # 'use' stubs out the default six callbacks that are required to run GenServer
+  #
+  # 1. `init/1` – Called when the server starts. Used to initialize state.
+  # 2. `handle_call/3` – Handles synchronous calls (`GenServer.call/2`).
+  # 3. `handle_cast/2` – Handles asynchronous messages (`GenServer.cast/2`).
+  # 4. `handle_info/2` – Handles all other messages (e.g. from `send/2` or process monitoring).
+  # 5. `terminate/2` – Called when the server is about to shut down.
+  # 6. `code_change/3` – Used for hot code upgrades.
+  #
+  use GenServer
+
+
+  # a default implementation. I'm guessing the version of elixir used now
+  # warns about missing this one out and urges you to add it in yourself.
+  def init(init_arg) do
+    {:ok, init_arg}
+  end
+
   def start do
     IO.puts "starting the pledge server..."
-    GenericServer.start(__MODULE__, [], @name)
+    GenServer.start(__MODULE__, [], name: @name)
   end
 
-  def create_pledge(name, amount), do: GenericServer.call(@name, {:create_pledge, name, amount})
+  def create_pledge(name, amount), do: GenServer.call(@name, {:create_pledge, name, amount})
 
-  def recent_pledges(), do: GenericServer.call(@name, :recent_pledges)
+  def recent_pledges(), do: GenServer.call(@name, :recent_pledges)
 
-  def total_pledged(), do: GenericServer.call(@name, :total_pledged)
+  def total_pledged(), do: GenServer.call(@name, :total_pledged)
 
-  def clear, do: GenericServer.cast(@name, :clear)
+  def clear, do: GenServer.cast(@name, :clear)
 
 
-  def handle_cast(message, _state), do: []
+  def handle_cast(:clear, _state), do: {:noreply, []}
 
-  def handle_call(:total_pledged, state) do
+  def handle_call(:total_pledged, _from, state) do
       total = Enum.map(state, &elem(&1, 1)) |> Enum.sum
-      {total, state}
+      {:reply, total, state}
   end
 
-  def handle_call(:recent_pledges, state), do: {state, state}
+  def handle_call(:recent_pledges, _from, state), do: {:reply, state, state}
 
-  def handle_call({:create_pledge, name, amount}, state) do
+  def handle_call({:create_pledge, name, amount}, _from, state) do
         # send to the external service
         {:ok, id} = send_pledge_to_service(name, amount)
 
@@ -75,7 +55,7 @@ defmodule Servy.PledgeServer do
         new_state = [ {name, amount} | most_recent_pledges]
 
         # return id
-        {id, new_state}
+        {:reply, id, new_state}
   end
 
   defp send_pledge_to_service(_name, _amount), do: {:ok, "pledge-#{:rand.uniform(1000)}"}
@@ -85,7 +65,7 @@ end
 
 alias Servy.PledgeServer
 
-pid = PledgeServer.start()
+{:ok, pid} = PledgeServer.start()
 
 send(pid, {:stop, "hammertime"})
 
@@ -94,7 +74,7 @@ IO.inspect PledgeServer.create_pledge("moe", 20)
 IO.inspect PledgeServer.create_pledge("curly", 30)
 IO.inspect PledgeServer.create_pledge("daisy", 40)
 
-IO.inspect PledgeServer.clear()
+# IO.inspect PledgeServer.clear()
 
 IO.inspect PledgeServer.create_pledge("grace", 50)
